// frontend/src/lib/api/models/ScheduleItemResponseDto.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type ScheduleItemResponseDto = {
    id: string;
    dayLabel: string;
    dayAccent: boolean;
    time: string;
    title: string;
    description?: string | null;
    sortOrder: number;
    isActive: boolean;
    createdAt: string;
    updatedAt: string;
};